uapi: convert list.h
